requirements.txt
langchain==0.0.167
openai==0.27.6
python-dotenv==1.0.0
tiktoken==0.4.0
transformers==4.29.1
telebot==0.0.5
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cohere==4.4.1
colorama==0.4.6
dataclasses-json==0.5.7
filelock==3.12.0
Flask==2.3.2
frozenlist==1.3.3
fsspec==2023.5.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.167
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pyparsing==3.0.9
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.13
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.29.1
typing-inspect==0.8.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
yarl==1.9.2